progress with big refactor
