+++ src/renderer.hpp
@@ -33,10 +33,8 @@ struct IterState<num_t,dims,Isaac<word_t,rparam>>
     typedef Point<num_t,dims> point_t;
     point_t p, t, v;
     Isaac<word_t,rparam>& rng;
-    const XForm<num_t,2,Isaac<word_t,rparam>> *xf;
     num_t *cw;
-    IterState(Isaac<word_t,rparam>& rng):
-        p(),t(),v(),rng(rng),xf(nullptr) {}
+    IterState(Isaac<word_t,rparam>& rng): p(),t(),v(),rng(rng) {}
     inline bool randBool()
     {
         return rng.next() & 1;
@@ -106,45 +104,36 @@ struct IterState<num_t,dims,Isaac<word_t,rparam>>
         return point_t(x);
     }
     // random point on the unit circle/sphere/hypersphere surface
-    ENABLE_IFEQ(dims,1,point_t) inline randDirection()
+    inline point_t randDirection()
     {
-        static const num_t table[2] = {-1.0,1.0};
-        return point_t(table[randBool()]);
-    }
-    ENABLE_IFEQ(dims,2,point_t) inline randDirection()
-    {
-        num_t a = (2.0*M_PI) * randNum();
-        num_t sa,ca;
-        sincosg(a,&sa,&ca);
-        return point_t(ca,sa);
-    }
-    ENABLE_IFEQ(dims,3,point_t) inline randDirection()
-    {
-        num_t p = acos(2.0*randNum()-1.0);
-        num_t t = (2.0*M_PI) * randNum();
-        num_t x = sin(t)*cos(p);
-        num_t y = sin(t)*sin(p);
-        num_t z = cos(t);
-        return point_t(x,y,z);
-    }
-    ENABLE_IF(dims>3,point_t) inline randDirection()
-    {
-        num_t x[dims];
-        num_t dummy;
-        for (size_t i = 0; i+1 < dims; i += 2)
-            randGaussianPair(x[i],x[i+1]);
-        if (dims % 2)
-            x[dims-1] = randGaussian();
-        point_t p(x);
-        return p/p.norm2();
-    }
-    inline const Affine<num_t,2>& getPreAffine() const
-    {
-        return xf->getPreAffine();
-    }
-    inline const Affine<num_t,2>& getPostAffine() const
-    {
-        return xf->getPostAffine();
+        if (dims == 1)
+            return point_t(copysign(1.0,randNum()-0.5));
+        else if (dims == 2)
+        {
+            num_t a = (2.0*M_PI) * randNum();
+            num_t sa,ca;
+            sincosg(a,&sa,&ca);
+            return point_t(ca,sa);
+        }
+        else if (dims == 3)
+        {
+            num_t p = acos(2.0*randNum()-1.0);
+            num_t t = (2.0*M_PI) * randNum();
+            num_t st,ct,sp,cp;
+            sincosg(p,&sp,&cp);
+            sincosg(t,&st,&ct);
+            return point_t(st*cp,st*sp,ct);
+        }
+        else
+        {
+            num_t x[dims];
+            for (size_t i = 0; i+1 < dims; i += 2)
+                randGaussianPair(x[i],x[i+1]);
+            if (dims % 2)
+                x[dims-1] = randGaussian();
+            point_t p(x);
+            return p/p.norm2();
+        }
     }
     // use cumulative weights to select xform
     inline u32 randXFormIndex()
@@ -178,7 +167,6 @@ private:
     Affine<num_t,dims> pre; // pre affine transformation
     Affine<num_t,dims> post; // post affine transformation
     bool has_pre,has_post;
-    u32 pc_flags; // precalculate flags
 public:
     XForm(){}
     // construct from JSON data
@@ -191,7 +179,6 @@ public:
             weight = 1.0; // unused
         if (weight <= 0.0)
             throw std::runtime_error("weights must be positive");
-        //num_t A[6];
         Json affine;
         has_pre = input.valueAt("pre_affine",affine);
         if (has_pre)
@@ -203,31 +190,21 @@ public:
             post = Affine<num_t,dims>(affine);
         else
             post = Affine<num_t,2>();
-        pc_flags = 0;
         for (Json varj : input["variations"].arrayValue())
         {
             XFormVar<num_t,dims,rand_t> var;
             std::string name = varj["name"].stringValue();
             const VarInfo<num_t,dims,rand_t>& varinfo =
-                //vars2d<num_t,rand_t>::data.at(name);
                 Variations<num_t,dims,rand_t>::get(name);
             var.func = varinfo.getFPtr();
             var.index = varp.size();
-            num_t weight = varj["weight"].floatValue();
-            if (weight == 0.0)
-                continue; // skip zero weight variations
             vars.push_back(var);
-            if (varinfo.getPPtr()) // store other parameters
-                varinfo.getPPtr()(*this,varj,weight,varp);
-            else // store just the weight by default
-                varp.push_back(weight);
-            pc_flags |= varinfo.getPCFlags();
+            varinfo.getPPtr()(varj,varp);
         }
     }
     // optimize xform
     void optimize()
     {
-        // zero weight variations are excluded in constructor
     }
     inline num_t getWeight() const
     {
@@ -250,28 +227,15 @@ public:
         return varp;
     }
     // iterate a state for the rendering process
-    ENABLE_IF(dims<=2,void) inline applyIteration(
-            IterState<num_t,dims,rand_t>& state) const
+    inline void applyIteration(IterState<num_t,dims,rand_t>& state) const
     {
         // for 2d, faster to apply identity affine than to branch
-        state.xf = this;
-        state.t = pre.apply_to(state.p);
-        state.v = Point<num_t,dims>();
-        for (XFormVar<num_t,dims,rand_t> v : vars)
-            v.func(state,varp.data()+v.index);
-        state.p = post.apply_to(state.v);
-    }
-    ENABLE_IF(dims>2,void) inline applyIteration(
-            IterState<num_t,dims,rand_t>& state) const
-    {
-        // for 3d, probably faster to skip identity affine
-        state.xf = this;
-        if (has_pre)
+        if (dims < 3 || has_pre)
             state.t = pre.apply_to(state.p);
         state.v = Point<num_t,dims>();
-        for (XFormVar<num_t,dims,rand_t> v : vars)
+        for (auto v : vars)
             v.func(state,varp.data()+v.index);
-        if (has_post)
+        if (dims < 3 || has_post)
             state.p = post.apply_to(state.v);
     }
 };
