+++ src/variations.hpp
@@ -1,5 +1,4 @@
 
 #pragma once
 
-#include "vars/vars.hpp"
-#include "vars/vars2d.hpp"
+#include "vars/vars_data.hpp"
